disasm: replace the legacy disassembler with a typed chunk model
The disassembler was rewritten from a stringly, generally typed version to a stricter, more data-oriented structure. The reader preserves byte-exact strings, uses typed IDs for constants, strings and protos. The new id data types reflect the actual numerical constraints, and the reader values are also verified more strictly for correctness.

The strings aren't cloned everywhere now, and are instead lazily resolved by carrying a Chunk reference and string ids.

Error handling by `anyhow` was introduced, dropping `thiserror` as everything drained down to a single eprintln anyways.

+++ src/common.rs
@@ -1,13 +1,13 @@
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct Spanned<T> {
     pub node: T,
-    pub pc: usize,
+    pub pc: u32,
 }
 
 impl<T> Spanned<T> {
     #[inline]
     #[must_use]
-    pub const fn new(node: T, pc: usize) -> Self {
+    pub const fn new(node: T, pc: u32) -> Self {
         Self { node, pc }
     }
 
@@ -28,7 +28,7 @@ impl<T> Spanned<T> {
 }
 
 pub trait ToSpanned {
-    fn to_spanned(self, pc: usize) -> Spanned<Self>
+    fn to_spanned(self, pc: u32) -> Spanned<Self>
     where
         Self: Sized,
     {
